+++ src/emulator/dataconnectEmulator.ts
@@ -1,4 +1,5 @@
 import * as childProcess from "child_process";
+import * as clc from "colorette";
 
 import { dataConnectLocalConnString } from "../api";
 import { Constants } from "./constants";
@@ -223,8 +224,9 @@ export class DataConnectEmulator implements EmulatorInstance {
   ): Promise<boolean> {
     const connectionString = localConnectionString ?? this.getLocalConectionString();
     if (!connectionString) {
-      this.logger.log("DEBUG", "No Postgres connection string found, not connecting to Postgres");
-      return false;
+      const msg = `No Postgres connection string found in '.firebaserc'. The Data Connect emulator will not be able to execute operations.
+Run ${clc.bold("firebase setup:emulators:dataconnect")} to set up a Postgres connection.`;
+      throw new FirebaseError(msg);
     }
     // The Data Connect emulator does not immediately start listening after started
     // so we retry this call with a brief backoff.